refactor(pack): refactor wifi config
