+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/modem/wifi/config/config.routing.js
@@ -1,0 +1,18 @@
+import controller from './config.controller';
+import template from './config.html';
+
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state('telecom.packs.pack.xdsl.line.modem.wifi', {
+    url: '/wifi',
+    views: {
+      'modemView@telecom.packs.pack.xdsl.line.modem': {
+        template,
+        controller,
+        controllerAs: 'ConfigWifiCtrl',
+      },
+    },
+    params: {
+      wifi: null,
+    },
+  });
+};
